1600 fix packaging
