+++ app/src/main/java/org/kiwix/kiwixmobile/di/ServiceScope.kt
@@ -15,20 +15,11 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
+package org.kiwix.kiwixmobile.di
 
-package org.kiwix.kiwixmobile.core.di.modules
+import javax.inject.Scope
+import kotlin.annotation.AnnotationRetention.RUNTIME
 
-import android.content.Context
-import android.location.LocationManager
-import dagger.Module
-import dagger.Provides
-import org.kiwix.kiwixmobile.core.di.KiwixScope
-
-@Module
-object KiwixModule {
-  @Provides
-  @KiwixScope
-  @JvmStatic
-  internal fun provideLocationManager(context: Context): LocationManager =
-    context.getSystemService(Context.LOCATION_SERVICE) as LocationManager
-}
+@Scope
+@Retention(RUNTIME)
+annotation class ServiceScope
