+++ app/src/main/java/org/kiwix/kiwixmobile/di/components/ServiceComponent.kt
@@ -16,19 +16,25 @@
  *
  */
 
-package org.kiwix.kiwixmobile.core.di.components
+package org.kiwix.kiwixmobile.di.components
 
-import dagger.Component
-import org.kiwix.kiwixmobile.core.di.KiwixScope
-import org.kiwix.kiwixmobile.core.di.modules.KiwixModule
-import org.kiwix.kiwixmobile.core.di.modules.ViewModelModule
+import android.app.Service
+import dagger.BindsInstance
+import dagger.Subcomponent
+import org.kiwix.kiwixmobile.di.ServiceScope
+import org.kiwix.kiwixmobile.di.modules.ServiceModule
+import org.kiwix.kiwixmobile.webserver.wifi_hotspot.HotspotService
 
-@KiwixScope
-@Component(
-  dependencies = [CoreComponent::class],
-  modules = [ViewModelModule::class, KiwixModule::class]
-)
-interface KiwixComponent {
-  fun activityComponentBuilder(): KiwixActivityComponent.Builder
-  fun serviceComponent(): ServiceComponent.Builder
+@Subcomponent(modules = [ServiceModule::class])
+@ServiceScope
+interface ServiceComponent {
+  fun inject(hotspotService: HotspotService)
+
+  @Subcomponent.Builder
+  interface Builder {
+
+    @BindsInstance fun service(service: Service): Builder
+
+    fun build(): ServiceComponent
+  }
 }
