+++ app/src/main/java/org/kiwix/kiwixmobile/di/modules/KiwixModule.kt
@@ -16,25 +16,19 @@
  *
  */
 
-package org.kiwix.kiwixmobile.core.di.components
+package org.kiwix.kiwixmobile.di.modules
 
-import android.app.Service
-import dagger.BindsInstance
-import dagger.Subcomponent
-import org.kiwix.kiwixmobile.core.di.ServiceScope
-import org.kiwix.kiwixmobile.core.di.modules.ServiceModule
-import org.kiwix.kiwixmobile.webserver.wifi_hotspot.HotspotService
+import android.content.Context
+import android.location.LocationManager
+import dagger.Module
+import dagger.Provides
+import org.kiwix.kiwixmobile.di.KiwixScope
 
-@Subcomponent(modules = [ServiceModule::class])
-@ServiceScope
-interface ServiceComponent {
-  fun inject(hotspotService: HotspotService)
-
-  @Subcomponent.Builder
-  interface Builder {
-
-    @BindsInstance fun service(service: Service): Builder
-
-    fun build(): ServiceComponent
-  }
+@Module
+object KiwixModule {
+  @Provides
+  @KiwixScope
+  @JvmStatic
+  internal fun provideLocationManager(context: Context): LocationManager =
+    context.getSystemService(Context.LOCATION_SERVICE) as LocationManager
 }
